Fix type error in usePushNotifications

+++ src/hooks/usePushNotifications.ts
@@ -1,5 +1,5 @@
 // src/hooks/usePushNotifications.ts
-'use client'; // Важно! Хук работает только на клиенте
+'use client';
 
 import { useState, useEffect, useCallback } from 'react';
 
@@ -13,41 +13,32 @@ interface PushSubscriptionJSON {
 }
 
 export function usePushNotifications() {
-  // Состояния хука
-  const [isSupported, setIsSupported] = useState(false);        // Поддерживает ли браузер
-  const [permission, setPermission] = useState<NotificationPermission>('default'); // Статус разрешения
-  const [subscription, setSubscription] = useState<PushSubscription | null>(null); // Текущая подписка
-  const [isLoading, setIsLoading] = useState(false);            // Загрузка
-  const [error, setError] = useState<string | null>(null);      // Ошибки
+  const [isSupported, setIsSupported] = useState(false);
+  const [permission, setPermission] = useState<NotificationPermission>('default');
+  const [subscription, setSubscription] = useState<PushSubscription | null>(null);
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
 
-  // Шаг 3.1: Проверка поддержки браузером
   useEffect(() => {
-    // Проверяем, что все необходимые API доступны в браузере
-    const supported = 'serviceWorker' in navigator &&           // Service Worker
-                     'PushManager' in window &&                  // Push API
-                     'Notification' in window;                    // Notifications API
+    const supported = 'serviceWorker' in navigator && 
+                     'PushManager' in window && 
+                     'Notification' in window;
     
     console.log('📱 Проверка поддержки PWA:', { supported });
     setIsSupported(supported);
 
     if (supported) {
-      // Если поддерживается, получаем текущий статус разрешений
       setPermission(Notification.permission);
-      
-      // Проверяем, есть ли уже подписка
       checkExistingSubscription();
     }
-  }, []); // Пустой массив - выполняется один раз при монтировании
+  }, []);
 
-  // Шаг 3.2: Проверка существующей подписки
   const checkExistingSubscription = async () => {
     try {
-      // Получаем регистрацию Service Worker
       const registration = await navigator.serviceWorker.getRegistration();
       console.log('🔍 Проверка Service Worker:', registration);
       
       if (registration) {
-        // Если есть Service Worker, проверяем наличие подписки
         const existingSubscription = await registration.pushManager.getSubscription();
         console.log('📦 Существующая подписка:', existingSubscription);
         setSubscription(existingSubscription);
@@ -57,7 +48,6 @@ export function usePushNotifications() {
     }
   };
 
-  // Шаг 3.3: Запрос разрешения на уведомления
   const requestPermission = useCallback(async (): Promise<boolean> => {
     if (!isSupported) {
       setError('Ваш браузер не поддерживает уведомления');
@@ -70,7 +60,6 @@ export function usePushNotifications() {
 
       console.log('🔔 Запрос разрешения на уведомления...');
       
-      // Запрашиваем разрешение у пользователя
       const result = await Notification.requestPermission();
       console.log('✅ Результат запроса:', result);
       
@@ -95,16 +84,11 @@ export function usePushNotifications() {
     }
   }, [isSupported]);
 
-  // Шаг 3.4: Вспомогательная функция для конвертации VAPID ключа
-  // VAPID ключ приходит в base64 формате, а Push API требует Uint8Array
   const urlBase64ToUint8Array = (base64String: string): Uint8Array => {
-    // Добавляем паддинг если нужно
     const padding = '='.repeat((4 - (base64String.length % 4)) % 4);
     const base64 = (base64String + padding)
-      .replace(/-/g, '+')  // Заменяем - на + (base64url -> base64)
-      .replace(/_/g, '/'); // Заменяем _ на /
-    
-    // Декодируем base64
+      .replace(/-/g, '+')
+      .replace(/_/g, '/');
     const rawData = window.atob(base64);
     const outputArray = new Uint8Array(rawData.length);
 
@@ -114,31 +98,25 @@ export function usePushNotifications() {
     return outputArray;
   };
 
-  // Шаг 3.5: Вспомогательная функция для преобразования подписки в JSON
-  // Нужно для отправки на сервер
   const subscriptionToJson = (subscription: PushSubscription): PushSubscriptionJSON => {
-    // Получаем ключи из подписки
     const key = subscription.getKey ? subscription.getKey('p256dh') : null;
     const auth = subscription.getKey ? subscription.getKey('auth') : null;
 
     return {
       endpoint: subscription.endpoint,
       keys: {
-        // Конвертируем ключи в base64 строки
         p256dh: key ? btoa(String.fromCharCode(...new Uint8Array(key))) : '',
         auth: auth ? btoa(String.fromCharCode(...new Uint8Array(auth))) : ''
       }
     };
   };
 
-  // Шаг 3.6: Подписка на уведомления
   const subscribe = useCallback(async () => {
     if (!isSupported) {
       setError('Ваш браузер не поддерживает уведомления');
       return null;
     }
 
-    // Если нет разрешения, запрашиваем его
     if (permission !== 'granted') {
       const granted = await requestPermission();
       if (!granted) return null;
@@ -150,15 +128,12 @@ export function usePushNotifications() {
 
       console.log('📡 Начинаем подписку на уведомления...');
 
-      // Шаг 3.6.1: Регистрируем Service Worker
       const registration = await navigator.serviceWorker.register('/notification-sw.js');
       console.log('✅ Service Worker зарегистрирован:', registration);
 
-      // Шаг 3.6.2: Ждем активации Service Worker
       await navigator.serviceWorker.ready;
       console.log('✅ Service Worker активирован');
 
-      // Шаг 3.6.3: Получаем VAPID ключ
       const vapidPublicKey = process.env.NEXT_PUBLIC_VAPID_PUBLIC_KEY;
       
       if (!vapidPublicKey) {
@@ -167,19 +142,17 @@ export function usePushNotifications() {
 
       console.log('🔑 VAPID ключ получен');
 
-      // Шаг 3.6.4: Конвертируем VAPID ключ
       const applicationServerKey = urlBase64ToUint8Array(vapidPublicKey);
 
-      // Шаг 3.6.5: Создаем подписку
+      // Исправление здесь - явное приведение типа
       const newSubscription = await registration.pushManager.subscribe({
-        userVisibleOnly: true, // Все уведомления будут видны пользователю
-        applicationServerKey: applicationServerKey
+        userVisibleOnly: true,
+        applicationServerKey: applicationServerKey as Uint8Array
       });
 
       console.log('✅ Подписка создана:', newSubscription);
       setSubscription(newSubscription);
 
-      // Шаг 3.6.6: Отправляем подписку на сервер
       const subscriptionData = subscriptionToJson(newSubscription);
       console.log('📤 Отправка подписки на сервер:', subscriptionData);
 
@@ -207,7 +180,6 @@ export function usePushNotifications() {
     }
   }, [isSupported, permission, requestPermission]);
 
-  // Шаг 3.7: Отписка от уведомлений
   const unsubscribe = useCallback(async () => {
     if (!subscription) {
       console.log('ℹ️ Нет активной подписки');
@@ -220,14 +192,12 @@ export function usePushNotifications() {
 
       console.log('🔇 Отписываемся от уведомлений...');
 
-      // Шаг 3.7.1: Удаляем подписку с сервера
       await fetch('/api/notifications/unsubscribe', {
         method: 'POST',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({ endpoint: subscription.endpoint })
       });
 
-      // Шаг 3.7.2: Отписываемся локально
       await subscription.unsubscribe();
       console.log('✅ Отписка выполнена');
       
@@ -244,15 +214,14 @@ export function usePushNotifications() {
     }
   }, [subscription]);
 
-  // Шаг 3.8: Возвращаем всё, что нужно компонентам
   return {
-    isSupported,      // Поддерживает ли браузер
-    permission,       // Текущий статус разрешения ('default', 'granted', 'denied')
-    subscription,     // Объект подписки (null если нет)
-    isLoading,        // Идёт ли загрузка
-    error,            // Сообщение об ошибке
-    requestPermission, // Функция запроса разрешения
-    subscribe,        // Функция подписки
-    unsubscribe       // Функция отписки
+    isSupported,
+    permission,
+    subscription,
+    isLoading,
+    error,
+    requestPermission,
+    subscribe,
+    unsubscribe
   };
 }
